Don't print strings, write bytes to stdout
This avoids the entire hassle with LC_CTYPE being set to something
that cannot encode Unicode. Thus, we're getting rid of the
UnicodeEncodeError and the trade-off is that our output is always
going to be UTF-8 encoded. But this is always the case! The HTML
template even has the UTF-8 meta data tag. Thus, when redirecting to a
file, you're always safe. The files will be UTF-8 encoded. And if your
terminal can handle UTF-8, you're also safe.