archive: handle incremental backup error
What happens when we get an empty page – i.e. we downloaded all the
pages, and yet we didn't find out last toot? This means that we must
have deleted it (or that the last toot was a boost of a toot that has
been deleted), or a bug. In either case, I don't think it's safe to
save this data. Abort! Abort!