Add error handling for URI without context
This commit adds an error message when the user attempts to fetch the
context of a toot that *is* in the archive but that does not contain any
other toots in its context (that is, it is not a reply to any toot in
the archive and no toot in the archive is a reply to it).

This commit does *not* return the toot itself, which could be helpful
but would be more verbose.