Change dismissing of other notifications
Instead of loading them all in order to count them, and then offering
a nice progress bar for their dismissal, dismiss them on the go
because it is taking so long. Use core.progress_bar, now: this at
least shows signs of life while going through the notifications. The
real problem, however, is that mastodon.notifications_dismiss seems to
have no effect, which is why it is taking longer and longer to go
through the list.