When downloading new statuses, insert them before existing ones
Currently, Mastodon Archive inserts statuses in reverse chronological
order on first load. But the next time you run it, it again downloads
your new posts in reverse chronological order, and appends them *at
the end* of the array of existing posts. This messes up the order in
Meow and in HTML export — if you had never split your archive, you
can generate the report and see that the order is wrong. E. g. if you
run it at 2021-08-01 for the first time, and then again at
2021-08-05, you'll get the following post order:
- 2021-08-01
- 2021-07-31
- 2021-07-30
- 2021-07-29
- 2021-08-05
- 2021-08-04
- 2021-08-03
- 2021-08-02

This commit fixes it.