Sort statuses after loading them from JSON
The previous commit fixes download order, but it doesn't help
existing archives. For those to show correctly in HTML export and
Meow, we sort user's own posts after loading them from JSON. We can't
do that for favourites, because data returned from API only contains
the favourited post publication date, but not the date at which the
user favourited it.