Fixing mentions expiration
We store the statuses mentioning us, but we can only dismiss the
notifications of these mentions. Previously, the code would attempt to
dismiss the notification using the id of the status and that won't
work. So now the flow is inverted: we're fetching all the
notifications and expiring them if they are old enough, it's either a
mention and we have it archived, or it's not a mention and
--delete-others was used.