Add --with-mutes, --with-blocks, --with-notes to archive command
Add the ability for the script to download the users you have muted,
the users you have blocked, and private notes you've added to your
follows, followers, mutes, and/or blocks.

It's currently not possible to download private notes added to _any_
users because the Mastodon API isn't capable of iterating them, so for
the time being this is the best we can do in terms of archiving
private notes. If/when the Mastodon API is enhanced to support
iterating through all private notes, this script can be updated to
take advantage of that.