Be smarter about video preloads
If a preview exists and it is different from the video source URL, use
it for the video poster attribute. In this case, no preloading
happens.

Alternatively, if the preview file is missing, or if the preview URL
is the same as the video source URL, as it seems to be in the Pleroma
example seen in GitHub issue #103, preload metadata and use no poster
image.