__init__.py needs to import core
The `main` function in `__init__.py` uses `core.all_accounts()` so
this file should imiport `core`. It was getting away with not doing it
because of the vagaries of how Python imports work with regards to
`core` being imported by other files that `__init__.py` imports, but
it shouldn't rely on that.