Don't keep trying to download files with permanent download errors
If we get 401 or 404 errors from both the local and remote URL for a
media attachment we are trying to download, create a flag file in the
media directory, the original file name with an added ".missing'
extension, and then try to don't download attachments for which the
flag file exists.

Furthermore, make the progress bar reflect the actual number of files
we are trying to download, with the ones we already have excluded, so
it is more accurate.